app/requirements.txt
flask
transformers
gtts
requests
python-dotenv
